policy: Move a few functions to policy_common
An upcoming commit will use the functions in
policy_common.
